Leave dependent files for packaging and add clear command in build.sh

+++ build.sh
@@ -1,26 +1,39 @@
 #!/bin/bash
 set -e
 
+function clear_build () {
+  rm -rf build dist *.egg-info
+}
+
 PYVER=$(python -c 'import sys; print(sys.version_info[0])')
 if [[ PYVER -ne 3 ]];then
   echo "The script should be run on python3"
   exit 1
 fi
 
+if [[ "$1" == "clear" ]]; then
+  clear_build
+  exit 0
+fi
+
 if [[ ("$1" = "test" && "$2" != "--ignore-test") || ("$1" = "build") || ("$1" = "deploy") ]]; then
   pip3 install -r requirements.txt
 
   VER=$(cat VERSION)
+  rm -rf ${VER}
   mkdir -p ${VER}
 
-  WGET_VER=$(curl http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/earlgrey/VERSION)
-  pip install --force-reinstall "http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/earlgrey/earlgrey-${WGET_VER}-py3-none-any.whl"
-  WGET_VER=$(curl http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/iconcommons/VERSION)
-  pip install --force-reinstall "http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/iconcommons/iconcommons-${WGET_VER}-py3-none-any.whl"
+  MOD_VER=$(curl http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/earlgrey/VERSION)
+  wget "http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/earlgrey/earlgrey-${MOD_VER}-py3-none-any.whl" -P ${VER}
+  pip install --force-reinstall "${VER}/earlgrey-${MOD_VER}-py3-none-any.whl"
+
+  MOD_VER=$(curl http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/iconcommons/VERSION)
+  wget "http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/iconcommons/iconcommons-${MOD_VER}-py3-none-any.whl" -P ${VER}
+  pip install --force-reinstall "${VER}/iconcommons-${MOD_VER}-py3-none-any.whl"
 
   if [[ -z "${ICONSERVICEPATH}" || ("$1" = "deploy") ]]; then
-    WGET_VER=$(curl http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/iconservice/VERSION)
-    pip install --force-reinstall "http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/iconservice/iconservice-${WGET_VER}-py3-none-any.whl"
+    wget "http://tbears.icon.foundation.s3-website.ap-northeast-2.amazonaws.com/${VER}/iconservice-${VER}-py3-none-any.whl" -P ${VER}
+    pip install --force-reinstall "${VER}/iconservice-${VER}-py3-none-any.whl"
   else
     if [ "$(pip3 list | grep iconservice)" ]; then
         pip uninstall iconservice -y
@@ -35,7 +48,7 @@ if [[ ("$1" = "test" && "$2" != "--ignore-test") || ("$1" = "build") || ("$1" = 
 
   if [ "$1" = "build" ] || [ "$1" = "deploy" ]; then
     pip install wheel
-    rm -rf build dist *.egg-info
+    clear_build
     python setup.py bdist_wheel
 
     if [ "$1" = "deploy" ]; then
@@ -48,6 +61,7 @@ if [[ ("$1" = "test" && "$2" != "--ignore-test") || ("$1" = "build") || ("$1" = 
 
       if [[ -z "${AWS_ACCESS_KEY_ID}" || -z "${AWS_SECRET_ACCESS_KEY}" ]]; then
         echo "Error: AWS keys should be in your environment"
+        rm -rf ${VER}
         exit 1
       fi
 
@@ -61,6 +75,7 @@ if [[ ("$1" = "test" && "$2" != "--ignore-test") || ("$1" = "build") || ("$1" = 
 else
   echo "Usage: build.sh [test|build|deploy]"
   echo "  test: run test"
+  echo "  clear: clear build and dist files"
   echo "  build: run test and build"
   echo "  build --ignore-test: run build"
   echo "  deploy: run test, build and deploy to s3"
